split up main_game
The module grew too large. Split it up.
